pyproject.toml
[project]
name = "aula_Django"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "isabellirosa", email = "isabelli.ifc@gmail.com"},
]
dependencies = [
    "django>=5.0.4",
    "djangorestframework>=3.15.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
